weatherapp/requirements.txt
Flask
pytest
requests
translate
Flask-Caching
gunicorn
python-dotenv

